+++ libs/pdf-generator/src/index.ts
@@ -12,3 +12,6 @@ export {
   pdfTable,
   pdfText,
 } from './lib/factories/block-factories';
+export * from './lib/models/pdf-generate-options';
+export * from './lib/models/pdf-result';
+export * from './generate-pdf';